Move folders to lib folder except hooks, components, contexts
