+++ lib/assets/index.ts
