+++ lib/constants/icon.ts
@@ -1,0 +1,4 @@
+export enum ICONS_LIST {
+  ARROW_DROP_DOWN = "arrow_drop_down",
+  ARROW_DROP_UP = "arrow_drop_up",
+}
