+++ lib/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./controller";
+export * from "./project";
