+++ lib/types/project.ts
@@ -1,0 +1,6 @@
+export interface project {
+  name: string;
+  description: string;
+  version: string;
+  routes: string[];
+}
